# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-офис собственника",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/files": {
      "post": {
        "tags": [
          "files"
        ],
        "summary": "Upload File",
        "operationId": "upload_file_api_v1_files_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_v1_files_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/knowledge/sources": {
      "get": {
        "tags": [
          "legal knowledge (local pilot)"
        ],
        "summary": "List Sources",
        "operationId": "list_sources_api_v1_knowledge_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceResponse"
                  },
                  "type": "array",
                  "title": "Response List Sources Api V1 Knowledge Sources Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "legal knowledge (local pilot)"
        ],
        "summary": "Ingest Source",
        "operationId": "ingest_source_api_v1_knowledge_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtifactResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "classification": {
            "type": "string",
            "title": "Classification"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "mime_type",
          "size_bytes",
          "sha256",
          "classification",
          "created_at"
        ],
        "title": "ArtifactResponse"
      },
      "Body_upload_file_api_v1_files_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_api_v1_files_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SourceCreate": {
        "properties": {
          "artifact_id": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Artifact Id"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "jurisdiction": {
            "type": "string",
            "pattern": "^[A-Z]{2}$",
            "title": "Jurisdiction"
          },
          "document_type": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Document Type"
          },
          "authority": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Authority"
          },
          "version": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Version"
          },
          "effective_from": {
            "type": "string",
            "format": "date",
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "APPROVED",
              "ACTIVE",
              "SUPERSEDED",
              "EXPIRED",
              "REVOKED"
            ],
            "title": "Status",
            "default": "DRAFT"
          },
          "classification": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "INTERNAL",
              "CONFIDENTIAL",
              "RESTRICTED"
            ],
            "title": "Classification",
            "default": "INTERNAL"
          },
          "language": {
            "type": "string",
            "pattern": "^[a-z]{2}$",
            "title": "Language",
            "default": "ru"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "artifact_id",
          "title",
          "jurisdiction",
          "document_type",
          "authority",
          "version",
          "effective_from"
        ],
        "title": "SourceCreate"
      },
      "SourceResponse": {
        "properties": {
          "artifact_id": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Artifact Id"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "jurisdiction": {
            "type": "string",
            "pattern": "^[A-Z]{2}$",
            "title": "Jurisdiction"
          },
          "document_type": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Document Type"
          },
          "authority": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Authority"
          },
          "version": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Version"
          },
          "effective_from": {
            "type": "string",
            "format": "date",
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "APPROVED",
              "ACTIVE",
              "SUPERSEDED",
              "EXPIRED",
              "REVOKED"
            ],
            "title": "Status",
            "default": "DRAFT"
          },
          "classification": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "INTERNAL",
              "CONFIDENTIAL",
              "RESTRICTED"
            ],
            "title": "Classification",
            "default": "INTERNAL"
          },
          "language": {
            "type": "string",
            "pattern": "^[a-z]{2}$",
            "title": "Language",
            "default": "ru"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "embedding_model": {
            "type": "string",
            "title": "Embedding Model"
          },
          "embedding_dimensions": {
            "type": "integer",
            "title": "Embedding Dimensions",
            "default": 128
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "reindex_of_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reindex Of Id"
          }
        },
        "type": "object",
        "required": [
          "artifact_id",
          "title",
          "jurisdiction",
          "document_type",
          "authority",
          "version",
          "effective_from",
          "id",
          "sha256",
          "embedding_model",
          "chunk_count"
        ],
        "title": "SourceResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}